# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Answer": {
        "description": "A language model's answer to a question about one route on one day, or why there is none.\n\n`text` is already filled in: the model wrote placeholders for every figure and the server put the\nengine's values there, after checking the rest carried no figure and no verdict. Absent unless\n`reason` is `ok`.",
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "title": "Citations",
            "type": "array"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "reason": {
            "enum": [
              "ok",
              "dropped",
              "off_topic",
              "disabled",
              "unavailable",
              "emergency"
            ],
            "title": "Reason",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "required": [
          "enabled",
          "reason",
          "citations"
        ],
        "title": "Answer",
        "type": "object"
      },
      "AskRequest": {
        "properties": {
          "history": {
            "items": {
              "$ref": "#/components/schemas/AskTurn"
            },
            "maxItems": 4,
            "title": "History",
            "type": "array"
          },
          "live": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LiveContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlanContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "question": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AskRequest",
        "type": "object"
      },
      "AskTurn": {
        "description": "One earlier exchange, as the client showed it (answers already filled in).",
        "properties": {
          "answer": {
            "maxLength": 2000,
            "title": "Answer",
            "type": "string"
          },
          "question": {
            "maxLength": 300,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question",
          "answer"
        ],
        "title": "AskTurn",
        "type": "object"
      },
      "AssessmentData": {
        "properties": {
          "alternatives": {
            "items": {
              "discriminator": {
                "mapping": {
                  "altRoute": "#/components/schemas/AltRoute",
                  "startEarlier": "#/components/schemas/StartEarlier"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/StartEarlier"
                },
                {
                  "$ref": "#/components/schemas/AltRoute"
                }
              ]
            },
            "title": "Alternatives",
            "type": "array"
          },
          "conditions": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/StopConditions"
              },
              "type": "array"
            },
            "title": "Conditions",
            "type": "object"
          },
          "forecast": {
            "$ref": "#/components/schemas/Forecast"
          },
          "gaps": {
            "items": {
              "enum": [
                "warnings",
                "snowline",
                "pace"
              ],
              "type": "string"
            },
            "title": "Gaps",
            "type": "array"
          },
          "hazards": {
            "items": {
              "$ref": "#/components/schemas/HazardDef"
            },
            "title": "Hazards",
            "type": "array"
          },
          "notEvaluated": {
            "items": {
              "$ref": "#/components/schemas/NotEvaluated"
            },
            "title": "Notevaluated",
            "type": "array"
          },
          "outcome": {
            "enum": [
              "assessed",
              "partial",
              "not_assessable"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "outcome",
          "stale",
          "forecast",
          "hazards",
          "gaps",
          "alternatives",
          "notEvaluated",
          "conditions"
        ],
        "title": "AssessmentData",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Leg": {
        "description": "A mapped section of the official route.",
        "properties": {
          "ascentM": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ascentm"
          },
          "cables": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cables"
          },
          "distanceKm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distancekm"
          },
          "fromIndex": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fromindex"
          },
          "fromStop": {
            "title": "Fromstop",
            "type": "string"
          },
          "grade": {
            "enum": [
              "T1",
              "T2",
              "T3",
              "T4",
              "T5",
              "T6"
            ],
            "title": "Grade",
            "type": "string"
          },
          "gradeEstimated": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gradeestimated"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "stopIds": {
            "items": {
              "type": "string"
            },
            "title": "Stopids",
            "type": "array"
          },
          "toIndex": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toindex"
          },
          "toStop": {
            "title": "Tostop",
            "type": "string"
          }
        },
        "required": [
          "id",
          "fromStop",
          "toStop",
          "stopIds",
          "grade"
        ],
        "title": "Leg",
        "type": "object"
      },
      "LiveContext": {
        "description": "Where the hiker is right now, during a hike, as field mode computes it.",
        "properties": {
          "cloud": {
            "anyOf": [
              {
                "enum": [
                  "above",
                  "touching",
                  "below"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud"
          },
          "eta": {
            "title": "Eta",
            "type": "integer"
          },
          "nextStopId": {
            "title": "Nextstopid",
            "type": "string"
          },
          "now": {
            "title": "Now",
            "type": "integer"
          },
          "offRoute": {
            "default": false,
            "title": "Offroute",
            "type": "boolean"
          },
          "remainingAscentM": {
            "title": "Remainingascentm",
            "type": "integer"
          },
          "remainingKm": {
            "title": "Remainingkm",
            "type": "number"
          },
          "status": {
            "enum": [
              "ahead",
              "behind",
              "pastCrux"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "now",
          "status",
          "nextStopId",
          "eta",
          "remainingKm",
          "remainingAscentM"
        ],
        "title": "LiveContext",
        "type": "object"
      },
...
      "Narration": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "hazards": {
            "items": {
              "$ref": "#/components/schemas/NarratedHazard"
            },
            "title": "Hazards",
            "type": "array"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "required": [
          "enabled",
          "hazards"
        ],
        "title": "Narration",
        "type": "object"
      },
      "NotEvaluated": {
        "properties": {
          "legIds": {
            "items": {
              "type": "string"
            },
            "title": "Legids",
            "type": "array"
          }
        },
        "required": [
          "legIds"
        ],
        "title": "NotEvaluated",
        "type": "object"
      },
...
      "PlaceRef": {
        "description": "A place the hiker picked out of search, as it comes back in.",
        "properties": {
          "latLng": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Latlng",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "latLng"
        ],
        "title": "PlaceRef",
        "type": "object"
      },
      "PlaceResult": {
        "description": "One search result. `rank` is swisstopo's own ordering; lower sorts first.\n\nNamed apart from `domain.PlaceHit` on purpose: that one is the parsed source object and must\nnever reach the wire, and `test_openapi_contract` checks by name that it does not.",
        "properties": {
          "latLng": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Latlng",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "latLng",
          "rank"
        ],
        "title": "PlaceResult",
        "type": "object"
      },
...
      "RetryResult": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "checkedAt": {
            "title": "Checkedat",
            "type": "integer"
          }
        },
        "required": [
          "available",
          "checkedAt"
        ],
        "title": "RetryResult",
        "type": "object"
      },
      "Route": {
        "properties": {
          "ascentM": {
            "title": "Ascentm",
            "type": "integer"
          },
          "bailoutName": {
            "title": "Bailoutname",
            "type": "string"
          },
          "bailoutStopId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bailoutstopid"
          },
          "cruxStopId": {
            "title": "Cruxstopid",
            "type": "string"
          },
          "descentM": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descentm"
          },
          "distanceKm": {
            "title": "Distancekm",
            "type": "number"
          },
          "elevations": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevations"
          },
          "fromName": {
            "title": "Fromname",
            "type": "string"
          },
          "geometry": {
            "anyOf": [
              {
                "items": {
                  "maxItems": 2,
                  "minItems": 2,
                  "prefixItems": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "number"
                    }
                  ],
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry"
          },
          "grade": {
            "enum": [
              "T1",
              "T2",
              "T3",
              "T4",
              "T5",
              "T6"
            ],
            "title": "Grade",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "legs": {
            "items": {
              "$ref": "#/components/schemas/Leg"
            },
            "title": "Legs",
            "type": "array"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/Stop"
            },
            "title": "Stops",
            "type": "array"
          },
          "toName": {
            "title": "Toname",
            "type": "string"
          },
          "turnaroundDefault": {
            "title": "Turnarounddefault",
            "type": "integer"
          },
          "waypoints": {
            "items": {
              "$ref": "#/components/schemas/Waypoint"
            },
            "title": "Waypoints",
            "type": "array"
          }
        },
        "required": [
          "id",
          "fromName",
          "toName",
          "grade",
          "distanceKm",
          "ascentM",
          "waypoints",
          "stops",
          "legs",
          "cruxStopId",
          "bailoutName",
          "turnaroundDefault"
        ],
        "title": "Route",
        "type": "object"
      },
      "RouteRequest": {
        "properties": {
          "from": {
            "$ref": "#/components/schemas/PlaceRef"
          },
          "to": {
            "$ref": "#/components/schemas/PlaceRef"
          },
          "via": {
            "items": {
              "$ref": "#/components/schemas/PlaceRef"
            },
            "title": "Via",
            "type": "array"
          }
        },
        "required": [
          "from",
          "to"
        ],
        "title": "RouteRequest",
        "type": "object"
      },
...
      "Stop": {
        "description": "A point on the day's timeline, in walking order (out and back).",
        "properties": {
          "breakMinutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breakminutes"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaceLabel"
              },
              {
                "$ref": "#/components/schemas/KeyLabel"
              }
            ],
            "title": "Label"
          },
          "legMinutes": {
            "title": "Legminutes",
            "type": "integer"
          },
          "waypointId": {
            "title": "Waypointid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "waypointId",
          "label",
          "legMinutes"
        ],
        "title": "Stop",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Waypoint": {
        "properties": {
          "elevationM": {
            "title": "Elevationm",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latLng": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Latlng",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "latLng",
          "elevationM"
        ],
        "title": "Waypoint",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Hiking safety assistant API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/forecast/retry": {
      "post": {
        "description": "\"Try again\" on the not-assessable screen: asks the forecast source whether it answers now.",
        "operationId": "retry_forecast_api_forecast_retry_post",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Forecast"
      }
    },
...
    "/api/routes": {
      "post": {
        "description": "Route between two searched places over the official trail network.",
        "operationId": "create_route_api_routes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Route"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Route"
      }
    },
    "/api/routes/search": {
      "get": {
        "description": "Place-name search, for picking the ends of a route.",
        "operationId": "search_places_api_routes_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaceResult"
                  },
                  "title": "Response Search Places Api Routes Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Places"
      }
    },
    "/api/routes/{route_id}": {
      "get": {
        "operationId": "get_route_api_routes__route_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "title": "Route Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Route"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Route"
      }
    },
    "/api/routes/{route_id}/ask": {
      "post": {
        "description": "A question about the hike on `date`, answered by a language model from the assessment and the plan.\n\nEvery figure in the answer is the engine's or the plan's, filled in by the server; the model decides\nnothing. `live` makes it about this minute of a hike under way. Never fails on the model's account:\n`reason` says why an answer has no text.",
        "operationId": "ask_about_route_api_routes__route_id__ask_post",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "title": "Route Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "en",
              "enum": [
                "en",
                "fr"
              ],
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Answer"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask About Route"
      }
    },
    "/api/routes/{route_id}/assessment": {
      "get": {
        "description": "Hazards for the hike on `date` (default: today in Switzerland), evaluated client-side at arrival.",
        "operationId": "get_route_assessment_api_routes__route_id__assessment_get",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "title": "Route Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentData"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Route Assessment"
      }
    },
    "/api/routes/{route_id}/narration": {
      "get": {
        "description": "The assessment's hazards phrased in `lang`, with the guidance each is grounded in.\n\nSeparate from the assessment so a slow model never holds it up. Never fails on the model's\naccount: a hazard without a `body` keeps its templated copy.",
        "operationId": "get_route_narration_api_routes__route_id__narration_get",
        "parameters": [
          {
            "in": "path",
            "name": "route_id",
            "required": true,
            "schema": {
              "title": "Route Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "en",
              "enum": [
                "en",
                "fr"
              ],
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Narration"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Route Narration"
      }
    }
  }
}